+++ src/test/converter2/issues/gh2336.ts
@@ -1,0 +1,13 @@
+export class ClassVersion {
+    /**
+     * Outer
+     * @returns Returns Outer
+     */
+    outer(): () => void {
+        /**
+         * Inner
+         * @returns Returns Inner
+         */
+        return () => {};
+    }
+}